rename XBUS.{c,h} to xbus.{c,h}
