+++ Makefile.common
@@ -28,7 +28,7 @@ SOURCES_C += \
 	$(4DO_DIR)/Madam.c \
 	$(4DO_DIR)/quarz.c \
 	$(4DO_DIR)/vdlp.c \
-	$(4DO_DIR)/XBUS.c \
+	$(4DO_DIR)/xbus.c \
 	$(4DO_DIR)/DiagPort.c \
 	$(4DO_DIR)/frame.c \
 	$(4DO_DIR)/bitop.c \
