+++ libfreedo/xbus.h
